requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<0.31
pydantic>=2.6,<3.0
python-dotenv>=1.0,<2.0
httpx>=0.27,<0.28
pytest>=8.2,<9.0
locust>=2.26,<3.0
